requirements.txt
langchain>=0.0.267
google-generativeai>=0.3.0
pandas>=2.0.0
numpy>=1.24.0
networkx>=3.1
python-dotenv>=1.0.0
streamlit>=1.26.0
python-docx>=0.8.11
matplotlib>=3.7.0
nltk>=3.8.1
scikit-learn>=1.2.0
